[ODS-5529] Nodes should share cached records when using an external cache (#552)

+++ Application/EdFi.Ods.Api/Caching/DescriptorsCache.cs
@@ -60,7 +60,7 @@ namespace EdFi.Ods.Api.Caching
             }
         }
 
-        private int InstanceId => _edFiOdsInstanceIdentificationProvider.GetInstanceIdentification();
+        private ulong InstanceId => _edFiOdsInstanceIdentificationProvider.GetInstanceIdentification();
 
         public string GetValue(string descriptorName, int id)
         {
